Move stepup files to appropriate namespace
The Stepup entities needed to be moved to a different namespace.
