+++ src/OpenConext/EngineBlock/Stepup/StepupGatewayLoaMapping.php
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-namespace OpenConext\EngineBlockBundle\Stepup;
+namespace OpenConext\EngineBlock\Stepup;
 
 use OpenConext\EngineBlock\Assert\Assertion;
 use OpenConext\EngineBlock\Exception\RuntimeException;
